requirements.txt
python-telegram-bot==20.3
apscheduler==3.10.0
sqlalchemy==2.0.20
python-dotenv==1.0.0
# cryptography==41.0.4
# pytest==7.4.0
# python-dateutil==2.8.2
# pywin32==306  # For Windows users, if you are on a different OS, you can remove this line
